pyproject.toml
[project]
name = "llamantin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "pydantic-settings>=2.6.1",
    "langchain-community>=0.3.8",
    "asyncio>=3.4.3",
    "langgraph>=0.2.53",
    "langchain>=0.3.8",
    "ollama>=0.4.1",
    "fastapi>=0.115.5",
    "pyside6>=6.8.0.2",
    "uvicorn>=0.32.1",
    "websockets>=14.1",
]

[dependency-groups]
dev = [
    "pytest-vcr>=1.0.2",
    "pytest>=8.3.3",
    "pytest-asyncio>=0.24.0",
]
